+++ src/input_context/input_condition/chord.rs
@@ -62,10 +62,7 @@ mod tests {
     use bevy_enhanced_input_macros::InputAction;
 
     use super::*;
-    use crate::{
-        action_value::ActionValueDim,
-        input_context::input_action::{ActionData, ActionsData},
-    };
+    use crate::input_context::input_action::{ActionData, ActionsData};
 
     #[test]
     fn chord() {